refactor : step4 제거 및 CarRacing 클래스 변경
