+++ src/main/java/car/Dice.java
@@ -1,4 +1,4 @@
-package step4;
+package car;
 
 import java.util.Random;
