+++ src/main/java/car/InputView.java
@@ -1,13 +1,22 @@
 package car;
 
+import java.util.ArrayList;
+import java.util.List;
 import java.util.Scanner;
 
 public class InputView {
     private final Scanner scanner = new Scanner(System.in);
 
-    public int carCountQuestion() {
-        System.out.println("자동차 대수는 몇 대 인가요?");
-        return scanner.nextInt();
+    public List<Car> carNameQuestion() throws Exception {
+        List<Car> cars = new ArrayList<>();
+
+        System.out.println("경주할 자동차 이름을 입력하새요(이름은 쉼표(,)를 긔준으로 구분).");
+        String answer = scanner.nextLine();
+        String[] carNames = answer.split(",");
+        for (String name : carNames) {
+            cars.add(new Car(name));
+        }
+        return cars;
     }
 
     public int attemptCountQuestion() {
